+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 * Don't execute lua onPaint callbacks when non rendering to improve fast-forward
 * We can duplicate multiple selected rows, and improve insertion/deletion
 * Remove backtrack savestate and invalidation
+* Pick a better state to rewind to
 
 ### Fixed